+++ saleor/product/migrations/0021_add_hstore_extension.py
@@ -3,9 +3,11 @@ from __future__ import unicode_literals
 from django.contrib.postgres.operations import HStoreExtension
 from django.db import migrations
 
+from saleor.core.hstore import h_store_extension
+
 
 class Migration(migrations.Migration):
 
     dependencies = [("product", "0020_attribute_data_to_class")]
 
-    operations = [HStoreExtension()]
+    operations = [h_store_extension]